LPS-171291 Move method DLAppServiceTestUtil.registerDLSyncEventProcessorMessageListener to class BaseDLAppWithRegisteringMessageListenerTestCase, prepare for next
