+++ modules/apps/document-library/document-library-test/src/testIntegration/java/com/liferay/document/library/app/service/test/DLAppServiceTestUtil.java
@@ -12,11 +12,6 @@ import com.liferay.document.library.kernel.service.DLAppServiceUtil;
 import com.liferay.document.library.test.util.BaseDLAppTestCase;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.configuration.test.util.ConfigurationTemporarySwapper;
-import com.liferay.portal.kernel.messaging.Destination;
-import com.liferay.portal.kernel.messaging.DestinationNames;
-import com.liferay.portal.kernel.messaging.Message;
-import com.liferay.portal.kernel.messaging.MessageBusUtil;
-import com.liferay.portal.kernel.messaging.MessageListener;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.search.Document;
 import com.liferay.portal.kernel.search.Field;
@@ -36,7 +31,6 @@ import com.liferay.portal.security.permission.DoAsUserThread;
 
 import java.util.Date;
 import java.util.Dictionary;
-import java.util.concurrent.atomic.AtomicInteger;
 
 import org.junit.Assert;
 
@@ -92,31 +86,6 @@ public class DLAppServiceTestUtil {
 		return new ConfigurationTemporarySwapper(
 			"com.liferay.document.library.configuration.DLConfiguration",
 			dictionary);
-	}
-
-	protected static AtomicInteger registerDLSyncEventProcessorMessageListener(
-		final String targetEvent) {
-
-		final AtomicInteger counter = new AtomicInteger();
-
-		Destination destination = MessageBusUtil.getDestination(
-			DestinationNames.DOCUMENT_LIBRARY_SYNC_EVENT_PROCESSOR);
-
-		destination.register(
-			new MessageListener() {
-
-				@Override
-				public void receive(Message message) {
-					Object event = message.get("event");
-
-					if (targetEvent.equals(event)) {
-						counter.incrementAndGet();
-					}
-				}
-
-			});
-
-		return counter;
 	}
 
 	protected static int runUserThreads(DoAsUserThread[] doAsUserThreads)
